UPDATE: Move all common methods from the test classes to a module base.py they will inherit from
